회원 수정 기능

+++ src/main/java/toyproject/board/controller/MemberController.java
@@ -13,8 +13,10 @@ import toyproject.board.domain.Member;
 import toyproject.board.domain.MemberDto;
 import toyproject.board.domain.Post;
 import toyproject.board.service.MemberService;
+import toyproject.board.service.PostService;
 
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -22,6 +24,7 @@ import java.util.List;
 public class MemberController {
 
     private final MemberService memberService;
+    private final PostService postService;
 
     @GetMapping("/signup")
     public String signup() {
@@ -38,7 +41,13 @@ public class MemberController {
                 .password(memberDto.getPassword())
                 .build();
 
-        memberService.signup(member);
+        Optional<Member> findMember = memberService.findByUsername(memberDto.getUsername());
+        if (findMember.isEmpty()) {
+            memberService.signup(member);
+        } else {
+            redirectAttributes.addAttribute("statusSignup", true);
+            return "redirect:/signup";
+        }
         redirectAttributes.addAttribute("status", true);
         return "redirect:/";
     }
@@ -50,7 +59,7 @@ public class MemberController {
 
     @PostMapping("/login")
     public String login(@ModelAttribute MemberDto memberDto, RedirectAttributes redirectAttributes) {
-        Member findMember = memberService.findByUsername(memberDto.getUsername());
+        Member findMember = memberService.findByUsername(memberDto.getUsername()).get();
 
         if (findMember.getPassword().equals(memberDto.getPassword())) {
             Long id = findMember.getId();
@@ -72,25 +81,45 @@ public class MemberController {
     }
 
     @GetMapping("/deleteMember/{id}")
-    public String DeleteForm(@PathVariable Long id, Model model) {
+    public String deleteForm(@PathVariable Long id, Model model) {
         Member member = memberService.findById(id);
         model.addAttribute("member", member);
-        return "MemberDelete";
+        return "deleteMember";
     }
 
     @PostMapping("/deleteMember/{id}") // form method post 설정안함
-    public String DeleteMember(@PathVariable Long id, @ModelAttribute MemberDto memberDto,
+    public String deleteMember(@PathVariable Long id, @ModelAttribute MemberDto memberDto,
                                RedirectAttributes redirectAttributes) {
         Member member = memberService.findById(id);
 
-        // String == 비교 말고 equals 사용해야함
+        // String == 비교 말고 equals 사용해야함. String 은 불변 객체
         if (member.getUsername().equals(memberDto.getUsername()) &&
-        member.getPassword().equals(memberDto.getPassword())) {
+                member.getPassword().equals(memberDto.getPassword())) {
             redirectAttributes.addAttribute("statusDeleteMember", true);
             memberService.deleteMember(member);
             return "redirect:/";
         }
         redirectAttributes.addAttribute("statusDeleteFail", true);
         return "redirect:/deleteMember/{id}";
+    }
+
+    @GetMapping("/updateMember/{id}")
+    public String updateMember(@PathVariable Long id, Model model) {
+        Member member = memberService.findById(id);
+        model.addAttribute("member", member);
+        return "updateMember";
+    }
+
+    @PostMapping("/updateMember/{id}")
+    public String updateMember(@PathVariable Long id, @ModelAttribute MemberDto memberDto,
+                               RedirectAttributes redirectAttributes) {
+
+        Member member = memberService.findById(id);
+        member.updateMember(memberDto.getUsername(), memberDto.getPassword());
+
+        memberService.signup(member);
+        redirectAttributes.addAttribute("statusUpdateMember", true);
+
+        return "redirect:/";
     }
 }
